search pagination

+++ database/migrations/2016_04_26_202253_create_vps_servers_table.php
@@ -27,7 +27,7 @@ class CreateVpsServersTable extends Migration
             $table->integer('hdd_size')->unsigned();
             $table->integer('cpu_version')->unsigned();
             $table->integer('cpu_cores')->unsigned();
-            $table->integer('cpu_power')->unsigned();
+            $table->float('cpu_power')->unsigned();
             $table->integer('bandwidth_tb')->unsigned();
             $table->integer('num_ips')->unsigned();
             $table->string('platform_os');
